+++ EF_Unity/Packages/cn.efefef.core/Runtime/Managers/Procedure/ProcedureEvents.cs
@@ -19,10 +19,10 @@ namespace EasyFramework.Managers.Procedure
     public struct ProcedureEnterEvent
     {
         /// <summary>新流程实例的唯一标识符</summary>
-        public uint Uid { get; }
+        public long Uid { get; }
 
         /// <summary>父流程实例的 Uid（0 表示根流程）</summary>
-        public uint ParentUid { get; }
+        public long ParentUid { get; }
 
         /// <summary>流程的具体类型</summary>
         public Type ProcedureType { get; }
@@ -30,7 +30,7 @@ namespace EasyFramework.Managers.Procedure
         /// <summary>当前嵌套深度（根深度为 1）</summary>
         public int Depth { get; }
 
-        public ProcedureEnterEvent(uint uid, uint parentUid, Type procedureType, int depth)
+        public ProcedureEnterEvent(long uid, long parentUid, Type procedureType, int depth)
         {
             Uid = uid;
             ParentUid = parentUid;
@@ -46,7 +46,7 @@ namespace EasyFramework.Managers.Procedure
     public struct ProcedureActivateEvent
     {
         /// <summary>被激活的流程实例 Uid</summary>
-        public uint Uid { get; }
+        public long Uid { get; }
 
         /// <summary>流程的具体类型</summary>
         public Type ProcedureType { get; }
@@ -54,7 +54,7 @@ namespace EasyFramework.Managers.Procedure
         /// <summary>当前嵌套深度</summary>
         public int Depth { get; }
 
-        public ProcedureActivateEvent(uint uid, Type procedureType, int depth)
+        public ProcedureActivateEvent(long uid, Type procedureType, int depth)
         {
             Uid = uid;
             ProcedureType = procedureType;
@@ -68,7 +68,7 @@ namespace EasyFramework.Managers.Procedure
     public struct ProcedureSuspendEvent
     {
         /// <summary>被挂起的流程实例 Uid</summary>
-        public uint Uid { get; }
+        public long Uid { get; }
 
         /// <summary>流程的具体类型</summary>
         public Type ProcedureType { get; }
@@ -76,7 +76,7 @@ namespace EasyFramework.Managers.Procedure
         /// <summary>当前嵌套深度</summary>
         public int Depth { get; }
 
-        public ProcedureSuspendEvent(uint uid, Type procedureType, int depth)
+        public ProcedureSuspendEvent(long uid, Type procedureType, int depth)
         {
             Uid = uid;
             ProcedureType = procedureType;
@@ -90,7 +90,7 @@ namespace EasyFramework.Managers.Procedure
     public struct ProcedureResumeEvent
     {
         /// <summary>被恢复的流程实例 Uid</summary>
-        public uint Uid { get; }
+        public long Uid { get; }
 
         /// <summary>流程的具体类型</summary>
         public Type ProcedureType { get; }
@@ -98,7 +98,7 @@ namespace EasyFramework.Managers.Procedure
         /// <summary>当前嵌套深度</summary>
         public int Depth { get; }
 
-        public ProcedureResumeEvent(uint uid, Type procedureType, int depth)
+        public ProcedureResumeEvent(long uid, Type procedureType, int depth)
         {
             Uid = uid;
             ProcedureType = procedureType;
@@ -112,7 +112,7 @@ namespace EasyFramework.Managers.Procedure
     public struct ProcedureExitEvent
     {
         /// <summary>正在退出的流程实例 Uid</summary>
-        public uint Uid { get; }
+        public long Uid { get; }
 
         /// <summary>流程的具体类型</summary>
         public Type ProcedureType { get; }
@@ -120,7 +120,7 @@ namespace EasyFramework.Managers.Procedure
         /// <summary>当前嵌套深度</summary>
         public int Depth { get; }
 
-        public ProcedureExitEvent(uint uid, Type procedureType, int depth)
+        public ProcedureExitEvent(long uid, Type procedureType, int depth)
         {
             Uid = uid;
             ProcedureType = procedureType;
@@ -134,7 +134,7 @@ namespace EasyFramework.Managers.Procedure
     public struct ProcedureTimeoutEvent
     {
         /// <summary>超时的流程实例 UID</summary>
-        public uint Uid { get; }
+        public long Uid { get; }
         
         /// <summary>流程的具体类型</summary>
         public Type ProcedureType { get; }
@@ -142,7 +142,7 @@ namespace EasyFramework.Managers.Procedure
         /// <summary>当前嵌套深度</summary>
         public int Depth { get; }
 
-        public ProcedureTimeoutEvent(uint uid, Type procedureType, int depth)
+        public ProcedureTimeoutEvent(long uid, Type procedureType, int depth)
         {
             Uid = uid;
             ProcedureType = procedureType;
